refactor logic for fetching moon data
