+++ scripts/file/index.js
@@ -1,5 +1,5 @@
 import { dirname, resolve } from "node:path";
-import { writeFileSync, readFileSync } from "node:fs";
+import { writeFileSync, readFileSync, existsSync, mkdirSync } from "node:fs";
 import { fileURLToPath } from "node:url";
 
 /**
@@ -20,9 +20,11 @@ export function relativePath(importMetaUrl, ...parts) {
  * @param {boolean} pretty true when data should be formatted before writing
  */
 export function writeJsonFileSync(filePath, data, pretty = false) {
+  ensureParentPath(filePath);
+
   const stringifyArgs = pretty ? [null, 2] : [];
   const stringified = JSON.stringify(data, ...stringifyArgs);
-  return writeFileSync(filePath, stringified);
+  return writeFileSync(filePath, stringified, { flag: "wx" });
 }
 
 /**
@@ -31,6 +33,13 @@ export function writeJsonFileSync(filePath, data, pretty = false) {
  * @returns {any} contents of file, parsed as JSON
  */
 export function readJsonFileSync(filePath) {
-  const data = fs.readFileSync(dataFilePath);
+  const data = readFileSync(filePath);
   return JSON.parse(data.toString());
+}
+
+function ensureParentPath(filePath) {
+  const parentPath = dirname(filePath);
+  if (!existsSync(parentPath)) {
+    mkdirSync(parentPath, { recursive: true });
+  }
 }
